+++ CHANGELOG.md
@@ -4,7 +4,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [0.3.1] - in development
+## [0.4.0] - in development
+
+### Changed
+
+- `ExtraWideUint` type moved from `PaillerParams` to `SchemeParams`. ([#205])
+
 
 ### Fixed